Solver: Refactor Validate.hs.
This just moves some functions and types into Validate.hs in preparation for
#4391.
