+++ cabal-install/Distribution/Solver/Modular/Assignment.hs
@@ -1,21 +1,16 @@
 module Distribution.Solver.Modular.Assignment
     ( Assignment(..)
+    , PAssignment
     , FAssignment
     , SAssignment
-    , PreAssignment(..)
-    , extend
     , toCPs
     ) where
 
-import Control.Applicative
-import Control.Monad
 import Data.Array as A
 import Data.List as L
 import Data.Map as M
 import Data.Maybe
 import Prelude hiding (pi)
-
-import Language.Haskell.Extension (Extension, Language)
 
 import Distribution.PackageDescription (FlagAssignment) -- from Cabal
 
@@ -29,72 +24,17 @@ import Distribution.Solver.Modular.Dependency
 import Distribution.Solver.Modular.Flag
 import Distribution.Solver.Modular.LabeledGraph
 import Distribution.Solver.Modular.Package
-import Distribution.Solver.Modular.Version
 
 -- | A (partial) package assignment. Qualified package names
 -- are associated with instances.
 type PAssignment    = Map QPN I
 
--- | A (partial) package preassignment. Qualified package names
--- are associated with constrained instances. Constrained instances
--- record constraints about the instances that can still be chosen,
--- and in the extreme case fix a concrete instance.
-type PPreAssignment = Map QPN (CI QPN)
 type FAssignment    = Map QFN Bool
 type SAssignment    = Map QSN Bool
 
 -- | A (partial) assignment of variables.
 data Assignment = A PAssignment FAssignment SAssignment
   deriving (Show, Eq)
-
--- | A preassignment comprises knowledge about variables, but not
--- necessarily fixed values.
-data PreAssignment = PA PPreAssignment FAssignment SAssignment
-
--- | Extend a package preassignment.
---
--- Takes the variable that causes the new constraints, a current preassignment
--- and a set of new dependency constraints.
---
--- We're trying to extend the preassignment with each dependency one by one.
--- Each dependency is for a particular variable. We check if we already have
--- constraints for that variable in the current preassignment. If so, we're
--- trying to merge the constraints.
---
--- Either returns a witness of the conflict that would arise during the merge,
--- or the successfully extended assignment.
-extend :: (Extension -> Bool) -- ^ is a given extension supported
-       -> (Language  -> Bool) -- ^ is a given language supported
-       -> (PkgconfigName -> VR  -> Bool) -- ^ is a given pkg-config requirement satisfiable
-       -> Var QPN
-       -> PPreAssignment -> [Dep QPN] -> Either (ConflictSet, [Dep QPN]) PPreAssignment
-extend extSupported langSupported pkgPresent var = foldM extendSingle
-  where
-
-    extendSingle :: PPreAssignment -> Dep QPN
-                 -> Either (ConflictSet, [Dep QPN]) PPreAssignment
-    extendSingle a (Ext  ext )  =
-      if extSupported  ext  then Right a
-                            else Left (varToConflictSet var, [Ext ext])
-    extendSingle a (Lang lang)  =
-      if langSupported lang then Right a
-                            else Left (varToConflictSet var, [Lang lang])
-    extendSingle a (Pkg pn vr)  =
-      if pkgPresent pn vr then Right a
-                          else Left (varToConflictSet var, [Pkg pn vr])
-    extendSingle a (Dep is_exe qpn ci) =
-      let ci' = M.findWithDefault (Constrained []) qpn a
-      in  case (\ x -> M.insert qpn x a) <$> merge ci' ci of
-            Left (c, (d, d')) -> Left  (c, L.map (Dep is_exe qpn) (simplify (P qpn) d d'))
-            Right x           -> Right x
-
-    -- We're trying to remove trivial elements of the conflict. If we're just
-    -- making a choice pkg == instance, and pkg => pkg == instance is a part
-    -- of the conflict, then this info is clear from the context and does not
-    -- have to be repeated.
-    simplify v (Fixed _ var') c | v == var && var' == var = [c]
-    simplify v c (Fixed _ var') | v == var && var' == var = [c]
-    simplify _ c              d                           = [c, d]
 
 -- | Delivers an ordered list of fully configured packages.
 --
